Create Hospital_Management_System.py
A medical facility needs a system that keeps a record of each patient. Each
record contains an integer identification of the patient and strings for
the date of the visit, the reason for the visit, and the treatment prescribed. The
system needs to provide three operations:
1. Search for a patient at a specific date and return the record of this
visit showing the reason for the visit and the treatment if found.
2. Find the total number of visits a specific patient made.
3. Add a new record to the system specifying: the patient id, date of
visit, reason of the visit, and the treatment.